requirements.txt
python-telegram-bot==13.15
Pillow
opencv-python
numpy
gradio_client
huggingface_hub
python-dotenv
